requirements.txt
boto3==1.28.59
botocore==1.31.59
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
jmespath==1.0.1
PyAudio==0.2.13
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
urllib3==1.26.17
